TINY-9302: Turn ticket-specific deletion and typing tests into webdriver tests
- use RealKeys to type to respect native browser behaviour
